Svelte => SvelteKit (#17)
* deleted old svelte files; initialized sveltekit project; moved api.js and store.js to src/lib; moved ui/public/assets to ui/static/assets

* moved components to src/lib

* src/App.svelte => src/routes/+page.svelte; fixed imports ()

* added /components

* src/routes/+layout.js: ssr = false (temporary fix for xterm not working in ssr context)

* moved Sidebar to +layout.svelte so that it's persistent across pages; made Sidebar a fixed element (fixed top-0 left-0 h-dvh z-10)

* force added api.js & store.js

---------
