+++ ui/.gitignore
@@ -1,2 +1,11 @@
+.DS_Store
 node_modules/
+/build
+/.svelte-kit
+/package
+.env
+.env.*
+!.env.example
+vite.config.js.timestamp-*
+vite.config.ts.timestamp-*
 pnpm-lock.yaml
