+++ ui/package.json
@@ -1,38 +1,26 @@
 {
-  "name": "devika-ui",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "name": "ui",
+  "version": "0.0.1",
+  "private": true,
   "scripts": {
-    "dev": "vite",
+    "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
+  "devDependencies": {
+    "@sveltejs/adapter-auto": "^3.0.0",
+    "@sveltejs/kit": "^2.0.0",
+    "@sveltejs/vite-plugin-svelte": "^3.0.0",
+    "autoprefixer": "^10.4.16",
+    "postcss": "^8.4.32",
+    "postcss-load-config": "^5.0.2",
+    "svelte": "^4.2.7",
+    "tailwindcss": "^3.3.6",
+    "vite": "^5.0.3"
+  },
+  "type": "module",
   "dependencies": {
-    "@babel/core": "^7.24.1",
-    "@babel/preset-env": "^7.24.1",
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-node-resolve": "^15.2.3",
-    "@sveltejs/vite-plugin-svelte": "^3.0.2",
-    "autoprefixer": "latest",
-    "css-loader": "^6.10.0",
-    "mini-css-extract-plugin": "^2.8.1",
-    "postcss": "latest",
-    "postcss-loader": "^8.1.1",
-    "rollup": "^4.13.0",
-    "rollup-plugin-css-only": "^4.5.2",
-    "rollup-plugin-livereload": "^2.0.5",
-    "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-svelte": "^7.2.0",
-    "rollup-plugin-terser": "^7.0.2",
-    "svelte": "^4.2.12",
-    "svelte-loader": "^3.2.0",
-    "tailwindcss": "latest",
-    "vite": "^5.1.6",
-    "xterm": "^5.2.0",
+    "xterm": "^5.3.0",
     "xterm-addon-fit": "^0.8.0"
   }
 }
