+++ ui/src/lib/components/Sidebar.svelte
@@ -1,5 +1,5 @@
 <div
-  class="flex flex-col w-20 bg-slate-950 p-4 space-y-4 items-center shadow-2xl shadow-indigo-700"
+  class="flex flex-col w-20 h-dvh fixed top-0 left-0 z-10 bg-slate-950 p-4 space-y-4 items-center shadow-2xl shadow-indigo-700"
 >
   <div class="p-5 rounded nav-button relative">
     <button class="text-gray-400 hover:text-white flex justify-center w-full">
