--- ui/src/main.js
@@ -1,8 +1,0 @@
-import App from './App.svelte';
-import './index.css';
-
-const app = new App({
-  target: document.getElementById('app'),
-});
-
-export default app;
