+++ ui/tailwind.config.cjs
@@ -1,0 +1,12 @@
+/** @type {import('tailwindcss').Config}*/
+const config = {
+  content: ["./src/**/*.{html,js,svelte,ts}"],
+
+  theme: {
+    extend: {},
+  },
+
+  plugins: [],
+};
+
+module.exports = config;
