--- ui/tailwind.config.js
@@ -1,11 +1,0 @@
-module.exports = {
-  purge: ["./index.html", "./src/**/*.{svelte,js,ts}"],
-  darkMode: false,
-  theme: {
-    extend: {},
-  },
-  variants: {
-    extend: {},
-  },
-  plugins: [],
-};
