--- ui/vite.config.mjs
@@ -1,9 +1,0 @@
-import { defineConfig } from 'vite';
-import { svelte } from '@sveltejs/vite-plugin-svelte';
-
-export default defineConfig({
-  plugins: [svelte()],
-  server: {
-    port: 3000,
-  },
-});
